+++ packages/twig-extension/TwigFunctions/TwigFunctionInterface.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Studiometa\Ui\TwigFunctions;
+
+interface TwigFunctionInterface
+{
+    /**
+     * Name of the function.
+     */
+    public function name(): string;
+
+    /**
+     * Get options for the TwigFunction instance.
+     */
+    public function options(): array;
+
+    /**
+     * Get the function callback.
+     */
+    public function callback(): callable;
+}