+++ tests/ExtensionTest.php
@@ -1,0 +1,5 @@
+<?php
+
+test('it renders an svg', function () {
+    assertTwigMatchesSnapshot("{{ meta_icon('mdi:alert') }}");
+});